Implement revised multi-element approach
